refactor(frontend): extract useCachedWorkspaceBootstrap hook
Four call sites duplicated the same cache-only TanStack Query observer for the
workspace bootstrap (users-tab, integrations-tab, stream-settings/members-tab,
bot-channels-section). Centralize the pattern in useCachedWorkspaceBootstrap
so each call site only states what it wants from the bootstrap.
